Blank field MessageBox and token encryption done.
Done with code that tells the user if any of the text fields are empty
and with the code that encrypts the API Token from the registrar.

Cleaned up QueryAPI.cs, removed code from old/abandoned ideas. Put in
dummy methods in QueryAPIName.cs so that it would compile for now.

Added string resources for error messages.