New cancel dialog.
Replaced the ugly WIX cancel dialog with a TaskDialog. This required
changing the Cancel button control on all dialogs in the installer.
I also wanted an icon so this commit includes added icon support to
TaskDialog.cs. Since I was adding icon support I went ahead and added an
error icon to the StatusDialog.

Localized the LICENSE file.