Added Open URL button when .NET 4.5 is missing.
On supported platforms without the .NET 4.5 framework installed, the
WelcomeDlg displays a message to the user with a URL of the .NET 4.5
installer. Before the user had to type it in manually since copy/paste
from the installer doesn't work, or do clickable links. I've added an
Open URL button at the bottom when this message appears that launches
the displayed link in the user's default browser instead of having them
type the URL or search for it.